requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
boto3==1.38.32
botocore==1.38.32
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
ecdsa==0.19.1
fastapi==0.115.12
frozenlist==1.6.2
gql==3.5.3
graphql-core==3.2.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jmespath==1.0.1
mangum==0.19.0
MarkupSafe==3.0.2
multidict==6.4.4
propcache==0.3.1
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-http-client==3.3.7
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.13.0
sendgrid==6.12.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
websockets==11.0.3
Werkzeug==3.1.3
yarl==1.20.0


